task-20: feat: вынести лидерборд в модули
