+++ src/css/main.scss
@@ -32,9 +32,7 @@
 // Модули
 @import '../modules/Auth/auth';
 @import '../modules/Profile/profile';
+@import '../modules/Leaderboard/leaderboard';
 
 // Лэйаут
 @import '../layout/Page/page';
-
-// Страницы
-@import '../pages/LeaderboardPage/leaderboard';
